# apps/bff/cmd/bff/main.go
//go:generate sh -c "cd ../.. && go run github.com/swaggo/swag/cmd/swag@v1.16.4 init -g cmd/bff/main.go -o docs --parseDependency && python3 scripts/postprocess_swagger.py --docs-dir docs"
// The global generator may expose unrelated baseline drift; LWC-174 validation asserts its own artifacts.

package main

import (
	"context"
	"errors"
	"fmt"
	"log"
	"net"
	"net/http"
	"os"
	"os/exec"
	"os/signal"
	"path/filepath"
	"strings"
	"syscall"
	"time"

	"github.com/gin-contrib/cors"
	"github.com/gin-gonic/gin"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"

	"github.com/rayer/llm-wiki-bff/internal/auth"
	"github.com/rayer/llm-wiki-bff/internal/buildinfo"
	conceptcache "github.com/rayer/llm-wiki-bff/internal/cache"
	"github.com/rayer/llm-wiki-bff/internal/config"
	"github.com/rayer/llm-wiki-bff/internal/exportjob"
	"github.com/rayer/llm-wiki-bff/internal/firestore"
	"github.com/rayer/llm-wiki-bff/internal/gcs"
	handlerraw "github.com/rayer/llm-wiki-bff/internal/handler"
	handlerv1 "github.com/rayer/llm-wiki-bff/internal/handler/v1"
	"github.com/rayer/llm-wiki-bff/internal/llm"
	"github.com/rayer/llm-wiki-bff/internal/localpipeline"
	"github.com/rayer/llm-wiki-bff/internal/middleware"
	"github.com/rayer/llm-wiki-bff/internal/observability"
	"github.com/rayer/llm-wiki-bff/internal/query"
	"github.com/rayer/llm-wiki-bff/internal/queryconfig"
	"github.com/rayer/llm-wiki-bff/internal/queryquality"
	"github.com/rayer/llm-wiki-bff/internal/queryruntime"
	"github.com/rayer/llm-wiki-bff/internal/search"
	store "github.com/rayer/llm-wiki-bff/internal/storage"
	"github.com/rayer/llm-wiki-bff/internal/syssettings"

	_ "github.com/rayer/llm-wiki-bff/docs" // swagger docs (generated by swag init)
)

//	@title			LLM Wiki BFF API
//	@version		1.0
//	@description	Backend-for-frontend API for the LLM Wiki project. Serves wiki sources and concepts with full-text search, import, and pipeline status.
//	@contact.name	LLM Wiki Team
//
//	@BasePath	/
//
//
//	@securityDefinitions.apikey	BearerAuth
//	@in							header
//	@name						Authorization
//	@description				JWT Bearer token issued by the standalone Auth service. Format: \"Bearer <token>\".
//
//	@securityDefinitions.apikey	ProjectHeader
//	@in							header
//	@name						X-Project-ID
//	@description				Project identifier header.

func main() {
	cfg, err := config.Load(".")
	if err != nil {
		log.Fatalf("Failed to load config: %v", err)
	}
	var stageConfig *queryconfig.Config
	if cfg.QueryStageConfigPath != "" {
		loaded, err := queryconfig.LoadFile(cfg.QueryStageConfigPath)
		if err != nil {
			log.Fatalf("Failed to load query stage config: %v", err)
		}
		stageConfig = &loaded
		log.Printf("Query stage config: schema=%d revision=%s digest=%s", loaded.SchemaVersion, loaded.ConfigRevision, loaded.ConfigDigest)
	}

	localMode := cfg.LocalCloudScope != ""
	if localMode && strings.TrimSpace(cfg.JWTSecret) == "" {
		log.Fatal("local cloud signing key is missing")
	}
	gcsClient, err := gcs.NewClient(cfg.Bucket)
	if err != nil {
		log.Fatalf("Failed to create GCS client: %v", err)
	}
	defer gcsClient.Close()
	wikiStore := store.RootStore(gcsClient)
	log.Printf("GCS target ready: gs://%s/%s", cfg.Bucket, cfg.LocalCloudScope)

	// Init Firestore client
	fsClient, err := firestore.NewClientWithDatabase(cfg.GCPProject, cfg.FirestoreDatabaseID, "", "")
	if err != nil {
		log.Fatalf("Failed to create Firestore client: %v", err)
	}
	defer fsClient.Close()
	if localMode {
		ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
		checkErr := errors.Join(fsClient.CheckAvailable(ctx), gcsClient.CheckAvailable(ctx))
		cancel()
		if checkErr != nil {
			log.Fatalf("local cloud target is unavailable: %v", checkErr)
		}
	}

	// Search metadata index
	idx := search.NewIndex()
	log.Printf("Search metadata index: not yet rebuilt — call /api/v1/pipeline/rebuild-index")
	log.Printf("Concept bodies: not yet rebuilt — call /api/v1/pipeline/rebuild-index")

	// Project-scoped concept cache for the V1 query path.
	conceptCache := conceptcache.New()
	log.Printf("Concept cache: not yet rebuilt — call /api/v1/pipeline/rebuild-index")

	productionExecutor, err := newProductionQueryExecutorWithStageConfig(cfg, conceptCache, stageConfig)
	if err != nil {
		log.Fatalf("Failed to create production query executor: %v", err)
	}
	build := buildinfo.Current()
	log.Printf("Build identity: commit=%s service=%s revision=%s", build.Commit, build.Service, build.Revision)

	// OpenTelemetry metrics (graceful fallback)
	provider, err := observability.InitMetrics(context.Background(), observabilityServiceName(os.Getenv("K_SERVICE")), observability.GetProjectID())
	if err != nil {
		log.Printf("[observability] WARNING: metrics init failed (continuing): %v", err)
	} else {
		defer func() {
			if err := provider.Shutdown(context.Background()); err != nil {
				log.Printf("[observability] metrics shutdown error: %v", err)
			}
		}()
	}

	// Handlers
	hV1 := handlerv1.New(wikiStore, fsClient, idx, conceptCache, nil, nil)
	hV1.SetQueryExecutor(productionExecutor)
	hV1.SetPipelineJobURL(cfg.PipelineJobURL)
	hV1.SetPipelineQuotaConfig(
		cfg.PipelineDailyLimit,
		cfg.PipelineCooldownSeconds,
		cfg.PipelineMinNewRaw,
		cfg.PipelineDemoUserIDs,
	)
	settingsStore := syssettings.NewStore(fsClient.Raw(), cfg.RegistrationEnabled)
	var nativeWorker *localpipeline.Manager
	if localMode {
		workerPath := strings.TrimSpace(os.Getenv("LOCAL_CLOUD_WORKER_PATH"))
		if workerPath == "" {
			log.Fatal("LOCAL_CLOUD_WORKER_PATH is required for local cloud pipeline execution")
		}
		workerPath, err = filepath.Abs(workerPath)
		if err != nil {
			log.Fatalf("resolve local worker path: %v", err)
		}
		workerInfo, statErr := os.Stat(workerPath)
		if statErr != nil || workerInfo.IsDir() || workerInfo.Mode()&0111 == 0 {
			log.Fatal("local cloud worker binary is unavailable or not executable")
		}
		if _, err := exec.LookPath(workerPath); err != nil {
			log.Fatalf("local cloud worker binary is not executable: %v", err)
		}
		nativeWorker, err = localpipeline.New(localpipeline.Config{
			Firestore: fsClient.Raw(), Storage: gcsClient, Worker: workerPath,
			Project: cfg.GCPProject, Bucket: cfg.Bucket, Database: cfg.FirestoreDatabaseID,
			Scope: cfg.LocalCloudScope, WorkDir: ".",
		})
		if err != nil {
			log.Fatalf("configure local pipeline executor: %v", err)
		}
		ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
		err = nativeWorker.ReconcileInterrupted(ctx)
		cancel()
		if err != nil {
			log.Fatalf("reconcile interrupted local pipeline records: %v", err)
		}
		hV1.SetLocalPipelineExecutor(nativeWorker)
	}

	r := newProductionRouter(cfg, localMode, gcsClient, fsClient, hV1, settingsStore, defaultRawScrapeHandlerFactory)

	listenHost := ":"
	if localMode {
		listenHost = "127.0.0.1:"
	}
	log.Printf("BFF listening on %s%s", listenHost, cfg.Port)
	log.Printf("Swagger UI: http://localhost:%s/swagger/index.html", cfg.Port)
	server := &http.Server{Addr: listenHost + cfg.Port, Handler: r, ReadHeaderTimeout: 10 * time.Second}
	serveErr := make(chan error, 1)
	go func() { serveErr <- server.ListenAndServe() }()
	stop, stopSignal := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGTERM)
	defer stopSignal()
	select {
	case err := <-serveErr:
		if err != nil && !errors.Is(err, http.ErrServerClosed) {
			log.Printf("BFF server stopped: %v", err)
		}
	case <-stop.Done():
	}
	if err := shutdownBFF(server, nativeWorker, 15*time.Second, 15*time.Second); err != nil {
		log.Printf("BFF shutdown: %v", err)
	}
}

type localPipelineCloser interface {
	Close(context.Context) error
}

func shutdownBFF(server *http.Server, nativeWorker localPipelineCloser, httpTimeout, workerTimeout time.Duration) error {
	var shutdownErrors []error
	httpCtx, cancelHTTP := context.WithTimeout(context.Background(), httpTimeout)
	if err := server.Shutdown(httpCtx); err != nil {
		shutdownErrors = append(shutdownErrors, fmt.Errorf("shut down HTTP server: %w", err))
		if closeErr := server.Close(); closeErr != nil {
			shutdownErrors = append(shutdownErrors, fmt.Errorf("force close HTTP server: %w", closeErr))
		}
	}
	cancelHTTP()
	if manager, ok := nativeWorker.(*localpipeline.Manager); ok && manager == nil {
		nativeWorker = nil
	}
	if nativeWorker != nil {
		workerCtx, cancelWorker := context.WithTimeout(context.Background(), workerTimeout)
		if err := nativeWorker.Close(workerCtx); err != nil {
			shutdownErrors = append(shutdownErrors, fmt.Errorf("stop local worker: %w", err))
		}
		cancelWorker()
	}
	return errors.Join(shutdownErrors...)
}

func newProductionQueryExecutor(cfg config.Config, conceptCache *conceptcache.Cache) (query.Executor, error) {
	var stageConfig *queryconfig.Config
	if cfg.QueryStageConfigPath != "" {
		loaded, err := queryconfig.LoadFile(cfg.QueryStageConfigPath)
		if err != nil {
			return nil, err
		}
		stageConfig = &loaded
	}
	return newProductionQueryExecutorWithStageConfig(cfg, conceptCache, stageConfig)
}

func newProductionQueryExecutorWithStageConfig(cfg config.Config, conceptCache *conceptcache.Cache, stageConfig *queryconfig.Config) (query.Executor, error) {
	if cfg.QueryStageConfigPath != "" && stageConfig == nil {
		return nil, errors.New("query stage config was not loaded")
	}
	if stageConfig != nil {
		sealed := *stageConfig
		expansionTemperature := sealed.Stages.QueryExpander.Temperature
		expansionClient := llm.NewClientWithOptions(cfg.DeepSeekAPIKey, llm.ClientOptions{
			Model: sealed.Stages.QueryExpander.Model, Temperature: &expansionTemperature,
			Reasoning: llm.Reasoning(sealed.Stages.QueryExpander.Reasoning),
		})
		if expansionClient == nil {
			return nil, errors.New("invalid sealed query expansion client configuration")
		}
		synthesisTemperature := sealed.Stages.AnswerSynthesizer.Temperature
		synthesisClient := llm.NewClientWithOptions(cfg.DeepSeekAPIKey, llm.ClientOptions{
			Model: sealed.Stages.AnswerSynthesizer.Model, Temperature: &synthesisTemperature,
			Reasoning: llm.Reasoning(sealed.Stages.AnswerSynthesizer.Reasoning),
		})
		if synthesisClient == nil {
			return nil, errors.New("invalid sealed query synthesis client configuration")
		}
		synthesisService := query.NewService(conceptCache, nil, synthesisClient)
		runtime, err := queryruntime.NewExecutor(sealed, conceptCache, expansionClient, nil, synthesisService)
		if err != nil {
			return nil, err
		}
		return runtime, nil
	}

	selectionLimit := cfg.QuerySelectionLimit
	if selectionLimit == 0 {
		selectionLimit = config.DefaultQuerySelectionLimit
	}
	explorationSlots := cfg.QuerySelectionExplorationSlots
	if cfg.QuerySelectionLimit == 0 && explorationSlots == 0 {
		explorationSlots = config.DefaultQuerySelectionExplorationSlots
	}
	evidenceThreshold := cfg.QuerySelectionEvidenceThreshold
	evidenceThresholdSet := cfg.QuerySelectionEvidenceThreshold != 0
	if evidenceThreshold == 0 {
		evidenceThreshold = config.DefaultQuerySelectionEvidenceThreshold
	}
	if cfg.QueryExpansionModel == "" {
		cfg.QueryExpansionModel = config.DefaultQueryExpansionModel
	}
	if cfg.AnswerSynthesisModel == "" {
		cfg.AnswerSynthesisModel = config.DefaultAnswerSynthesisModel
	}
	if cfg.AnswerSynthesisReasoning == "" {
		cfg.AnswerSynthesisReasoning = config.DefaultAnswerSynthesisReasoning
	}
	if cfg.QueryExpansionModel != "" && cfg.QueryExpansionModel != config.DefaultQueryExpansionModel {
		return nil, fmt.Errorf("query expansion model must be %s", config.DefaultQueryExpansionModel)
	}
	synthesisClient := llm.NewClientWithOptions(cfg.DeepSeekAPIKey, llm.ClientOptions{Model: cfg.AnswerSynthesisModel, Reasoning: cfg.AnswerSynthesisReasoning})
	if cfg.DeepSeekAPIKey != "" && synthesisClient == nil {
		return nil, errors.New("invalid answer synthesis client configuration")
	}
	temperature := 0.0
	expansionClient := llm.NewClientWithOptions(cfg.DeepSeekAPIKey, llm.ClientOptions{Model: cfg.QueryExpansionModel, Temperature: &temperature, Reasoning: config.DefaultQueryExpansionReasoning})
	var expansionProvider queryquality.ChatProvider
	if expansionClient != nil {
		expansionProvider = expansionClient
	}
	legacyExpander, err := llm.NewExpander(nil, "lifestyle")
	if err != nil {
		return nil, err
	}
	legacy := query.NewService(conceptCache, legacyExpander, synthesisClient)
	return queryquality.NewProductionExecutor(conceptCache, expansionProvider, legacy, legacy, queryquality.Options{
		SelectionLimit: selectionLimit, ExplorationSlots: explorationSlots,
		EvidenceThreshold: evidenceThreshold, EvidenceThresholdSet: evidenceThresholdSet,
		KeywordsPerAttempt:    cfg.QueryExpansionKeywordsPerAttempt,
		ExpansionAttempts:     cfg.QueryExpansionAttempts,
		RareDocumentFrequency: cfg.QueryMatchingRareKeywordMaxDocumentFrequency,
	})
}

type rawScrapeHandler interface {
	ScrapeRaw(*gin.Context)
}

type rawScrapeHandlerFactory func(scopedClient *gcs.Client, fsClient *firestore.Client) rawScrapeHandler

func defaultRawScrapeHandlerFactory(scopedClient *gcs.Client, fsClient *firestore.Client) rawScrapeHandler {
	return handlerraw.New(scopedClient, fsClient, nil, nil, nil)
}

func localAuthHostAllowlist(cfg config.Config) gin.HandlerFunc {
	allowed := make(map[string]struct{})
	for _, host := range cfg.AllowedHostsFor(true) {
		allowed[strings.ToLower(strings.TrimSpace(host))] = struct{}{}
	}
	return func(c *gin.Context) {
		host := strings.ToLower(strings.TrimSpace(c.Request.Host))
		if parsed, _, err := net.SplitHostPort(host); err == nil {
			host = parsed
		}
		if _, ok := allowed[host]; !ok {
			c.AbortWithStatusJSON(http.StatusBadRequest, gin.H{"error": "invalid Host header"})
			return
		}
		c.Next()
	}
}

func newProductionRouter(
	cfg config.Config,
	localMode bool,
	gcsClient *gcs.Client,
	fsClient *firestore.Client,
	hV1 *handlerv1.Handler,
	settingsStore syssettings.RegistrationGate,
	rawScrapeFactory rawScrapeHandlerFactory,
) *gin.Engine {
	r := gin.Default()
	r.Use(middleware.SecurityHeaders(true))

	// OpenTelemetry latency middleware (per-endpoint)
	r.Use(middleware.LatencyMiddleware())

	// CORS — must be BEFORE routes
	allowOrigins := cfg.AllowedOriginsFor(localMode)
	r.Use(cors.New(cors.Config{
		AllowOrigins:     allowOrigins,
		AllowMethods:     []string{"GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS"},
		AllowHeaders:     []string{"Content-Type", "Authorization", "X-Project-ID", "Idempotency-Key"},
		AllowCredentials: true,
	}))

	registerPublicRoutes(r, settingsStore, cfg.AuthServiceURL, !localMode)
	r.POST("/internal/profile/dispatch", productionProfileDispatcher(hV1, cfg).Gin)
	r.GET("/api/v1/query/config", hV1.QueryConfig)

	// Temporary Stage A compatibility lane for the frontend Auth cutover.
	var sessions *auth.RefreshSessionAuthority
	authRoutes := r.Group("/api/v1/auth")
	if localMode {
		authRoutes.Use(localAuthHostAllowlist(cfg))
	}
	authRoutes.Use(auth.CompatibilityBodyLimit())
	cookiePolicy := auth.LegacyRefreshCookiePolicy()
	if localMode {
		cookiePolicy = auth.LocalRefreshCookiePolicy()
	}
	if fsClient == nil || fsClient.Raw() == nil {
		unavailable := func(c *gin.Context) {
			c.JSON(http.StatusServiceUnavailable, gin.H{"error": "auth routes require Firestore"})
		}
		authRoutes.POST("/login", unavailable)
		authRoutes.POST("/register", unavailable)
		authRoutes.POST("/refresh", unavailable)
		authRoutes.POST("/logout", auth.LogoutHandlerWithCookiePolicy(cookiePolicy))
	} else {
		identityRepository := auth.NewIdentityRepository(fsClient.Raw())
		sessionEnvironment := strings.TrimSpace(cfg.AuthSessionEnvironment)
		if sessionEnvironment == "" {
			sessionEnvironment = strings.TrimSpace(cfg.FirestoreDatabaseID)
			if sessionEnvironment == "" {
				sessionEnvironment = "default"
			}
		}
		sessions = auth.NewRefreshSessionAuthorityWithConfig(fsClient.Raw(), auth.SessionAuthorityConfig{
			Environment: sessionEnvironment, Migration: auth.RefreshSessionMigrationMode(cfg.AuthSessionMigration),
		})
		authRoutes.POST("/login", middleware.NewRateLimiter(10, time.Minute), auth.LoginHandlerWithRepositoryAndSessionAuthority(identityRepository, cfg.JWTSecret, cookiePolicy, sessions))
		authRoutes.POST("/register", middleware.NewRateLimiter(5, time.Minute), auth.RegisterHandlerWithRepository(identityRepository, cfg.JWTSecret, settingsStore))
		authRoutes.POST("/refresh", auth.RefreshHandlerWithSessionAuthority(sessions, cfg.JWTSecret, cookiePolicy))
		authRoutes.POST("/logout", auth.LogoutHandlerWithSessionAuthority(sessions, cfg.JWTSecret, cookiePolicy))
	}
	if fsClient == nil {
		wireCLIAuthorities(hV1, nil, sessions)
	} else {
		wireCLIAuthorities(hV1, fsClient.Raw(), sessions)
	}

	// ── Swagger UI ──
	r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))

	// ── API v1 (with JWT auth + project routing) ──
	v1 := r.Group("/api/v1")

	v1.Use(hV1.AccountAuth(cfg))
	{
		v1.GET("/health", hV1.Health)
		v1.GET("/ready", hV1.Ready)
		v1.GET("/projects", hV1.ListProjects)
		v1.POST("/init-project", hV1.InitProject)
		v1.PATCH("/projects/:projectID", hV1.RenameProject)
		v1.GET("/projects/:pid/status", hV1.ProjectStatus)
	}

	registerAdminRoutes(r, cfg, hV1, settingsStore)

	v1.Use(auth.ProjectMiddleware())
	{
		v1.POST("/pipeline/rebuild-index", hV1.RebuildIndex)
		v1.GET("/projects/:pid/recompile-all/capability", hV1.RecompileAllCapability)
		v1.POST("/projects/:pid/recompile-all", hV1.RecompileAll)
		profileRoutes := v1.Group("")
		profileRoutes.Use(hV1.TrialDemoAccessGuard())
		profileRoutes.GET("/projects/:pid/profile", hV1.GetProfile)
		profileRoutes.PUT("/projects/:pid/profile", hV1.PutProfile)
		profileRoutes.GET("/projects/:pid/profile/bootstrap-guidance", hV1.GetProfileBootstrapGuidance)
		profileRoutes.GET("/projects/:pid/profile/guidance/:revision", hV1.GetProfileGuidanceArtifact)
		profileRoutes.POST("/projects/:pid/profile/bootstrap-guidance/:revision/confirm", hV1.ConfirmProfileBootstrapGuidance)
		profileRoutes.POST("/projects/:pid/profile/candidates/:candidateID/confirm", hV1.ConfirmProfileCandidate)
		profileRoutes.POST("/projects/:pid/profile/candidates/:candidateID/retry", hV1.RetryProfileCandidate)
		profileRoutes.POST("/projects/:pid/profile/derivation/retry", hV1.RetryProfileDerivation)
		profileRoutes.GET("/projects/:pid/profile/jobs/:jobID", hV1.GetProfileJob)
		var exportRepo exportjob.Repository
		var projectVerifier exportjob.ProjectVerifier
		var archiveStore *exportjob.CloudArchiveStore
		if fsClient != nil && fsClient.Raw() != nil {
			exportRepo = exportjob.NewFirestoreRepository(fsClient.Raw())
			projectVerifier = exportjob.NewFirestoreProjectVerifier(fsClient.Raw())
		}
		if !localMode && cfg.Bucket != "" && cfg.ExportSigningServiceAccount != "" {
			var err error
			archiveStore, err = exportjob.NewCloudArchiveStore(context.Background(), cfg.Bucket, cfg.ExportSigningServiceAccount)
			if err != nil {
				log.Printf("WARNING: export storage unavailable: %v", err)
			}
		}
		exportHandler := newExportHTTPHandler(localMode, cfg, exportRepo, projectVerifier, archiveStore)
		exportRoutes := v1.Group("")
		exportRoutes.Use(hV1.TrialDemoAccessGuard())
		exportHandler.Register(exportRoutes)

		v1.GET("/index", hV1.Index)
		v1.POST("/query", hV1.Query)
		v1.GET("/sources", hV1.ListSources)
		v1.GET("/sources/:id/annotation", hV1.GetAnnotation)
		v1.PUT("/sources/:id/annotation", hV1.PutAnnotation)
		v1.GET("/sources/:id", hV1.GetSource)
		v1.GET("/concepts", hV1.ListConcepts)
		v1.GET("/concepts/:id", hV1.GetConcept)
		v1.POST("/import", hV1.Import)
		v1.POST("/pipeline/run", hV1.PipelineRun)
		v1.GET("/pipeline/status", hV1.PipelineStatus)
		v1.GET("/pipeline/log", hV1.PipelineLog)
		registerRawScrapeRoute(v1, gcsClient, fsClient, rawScrapeFactory)
		v1.GET("/raw", hV1.RawList)
		v1.GET("/raw/:filename", hV1.RawPreview)
		v1.POST("/raw/upload", hV1.RawUpload)
		v1.GET("/status", hV1.Status)
		v1.GET("/metrics", hV1.PrometheusMetrics)
	}

	// ── Static frontend ──
	r.StaticFile("/", "./static/index.html")
	r.Static("/static", "./static")
	r.NoRoute(func(c *gin.Context) {
		// SPA fallback: serve index.html for non-API routes
		if !strings.HasPrefix(c.Request.URL.Path, "/api") {
			c.File("./static/index.html")
		}
	})

	return r
}

func newExportHTTPHandler(localMode bool, cfg config.Config, repo exportjob.Repository, projects exportjob.ProjectVerifier, cloudArchives *exportjob.CloudArchiveStore) *exportjob.HTTPHandler {
	var archives exportjob.ArchiveStore
	var starter exportjob.JobStarter
	if !localMode {
		if cloudArchives != nil {
			archives = cloudArchives
		}
		starter = exportjob.NewCloudRunJobStarter(cfg.ExportJobURL)
	}
	return exportjob.NewHTTPHandler(repo, projects, archives, starter)
}

func registerPublicRoutes(r *gin.Engine, settingsStore syssettings.RegistrationGate, authServiceURL string, demoEnabled bool) {
	// Public routes are registered on the root router, outside JWT and project middleware.
	_ = authServiceURL
	r.GET("/api/v1/public/config", syssettings.PublicConfigHandler(settingsStore, demoEnabled))
	r.GET("/api/v1/public/version", buildinfo.Handler())
}

const legacyObservabilityServiceName = "llm-wiki-bff-dev"

func observabilityServiceName(kService string) string {
	if serviceName := strings.TrimSpace(kService); serviceName != "" {
		return serviceName
	}
	return legacyObservabilityServiceName
}

func registerAdminRoutes(r *gin.Engine, cfg config.Config, hV1 *handlerv1.Handler, settingsStore syssettings.RegistrationGate) {
	admin := r.Group("/api/v1/admin")
	admin.Use(hV1.AccountAuth(cfg), auth.AdminOnly())
	{
		admin.GET("/settings", syssettings.AdminGetSettingsHandler(settingsStore))
		admin.PATCH("/settings", syssettings.AdminPatchSettingsHandler(settingsStore))
		admin.POST("/settings/announcement/publish", syssettings.AdminPublishAnnouncementHandler(settingsStore))
		admin.GET("/projects", hV1.AdminProjects)
		admin.PATCH("/projects/:id", hV1.AdminRenameProject)
		admin.DELETE("/projects/:id", hV1.AdminDeleteProject)
		admin.POST("/projects/:id/pipeline", hV1.AdminPipelineTrigger)
		admin.GET("/projects/:id/pipeline/status", hV1.AdminPipelineStatus)
		admin.POST("/projects/:id/rebuild-index", hV1.AdminRebuildIndex)
		admin.GET("/users", hV1.AdminListUsers)
		admin.PATCH("/users/:id", hV1.AdminUpdateUser)
		admin.DELETE("/users/:id", hV1.AdminDeleteUser)
	}
}

func registerRawScrapeRoute(v1 *gin.RouterGroup, gcsClient *gcs.Client, fsClient *firestore.Client, rawScrapeFactory rawScrapeHandlerFactory) {
	if rawScrapeFactory == nil {
		rawScrapeFactory = defaultRawScrapeHandlerFactory
	}
	v1.POST("/raw/scrape", func(c *gin.Context) {
		if gcsClient == nil {
			c.JSON(http.StatusServiceUnavailable, handlerraw.ErrorResponse{Error: "storage client is not configured"})
			return
		}

		userID := strings.TrimSpace(c.GetString("userID"))
		projectID := strings.TrimSpace(c.GetString("projectID"))
		if userID == "" {
			c.JSON(http.StatusUnauthorized, handlerraw.ErrorResponse{Error: "user not authenticated"})
			return
		}
		if projectID == "" {
			c.JSON(http.StatusBadRequest, handlerraw.ErrorResponse{Error: "project ID is required"})
			return
		}

		rawScrapeHandler := rawScrapeFactory(gcsClient.WithScope(userID, projectID), fsClient)
		rawScrapeHandler.ScrapeRaw(c)
	})
}